pyproject.toml
[tool.poetry]
name = "alpha-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "alpha_store"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
alembic = "^1.9.3"
coverage = "^7.1.0"
boto3 = "^1.26.67"
waitress = "^2.1.2"
pg8000 = "^1.29.4"
prospector = "^1.8.4"
loguru = "^0.6.0"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
requests = "^2.28.2"
parameterized = "^0.8.1"
flask-marshmallow = "^0.14.0"
pandas = "^1.5.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.21.1"
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
